-Copied parts out to separate files. -Still need to handle UDT's with other UDT as member variables
