+++ Oscillator.cpp
@@ -1,0 +1,57 @@
+#include "Oscillator.h"
+
+Oscillator::Oscillator()
+{
+    semitone  = 0;
+    waveform = 'N';
+    phaseIndex = 0.f;
+    noiseLevel = -100.f;
+    lfoAmount = 0.f;
+}
+Oscillator::~Oscillator()
+{
+    std::cout << "Osc last phase is " << phaseIndex << std::endl;
+}
+
+float Oscillator::outputLFOSignal(char lfoWave, float amplitudeLevel)
+{
+    std::cout << "LFO Amnt before: " << lfoAmount << std::endl;
+
+    for(int i = 0; i < 2; ++i )
+    {
+        waveform = lfoWave;
+        amplitudeLevel = .75f * i;
+    }
+
+    return amplitudeLevel;
+}
+
+void Oscillator::syncOscillator(Oscillator oscToSyncTo)
+{
+    phaseIndex = oscToSyncTo.phaseIndex;
+}
+
+int Oscillator::nextSamp(int freq, int phase, char whichWave)
+{
+    int sampValue = 0;
+    if(whichWave == 'N')
+    {
+        sampValue = freq * (100 - (phase % 200));
+    }
+    return sampValue;
+}
+
+void Oscillator::outputSignal(int frequency, int bufferSize)
+{
+    std::cout << "Value of sample at " ;
+    for (int bufferSample = 0; bufferSample < bufferSize; ++bufferSample)
+    {
+        std::cout << bufferSample << " = " << nextSamp(frequency, bufferSample, waveform) << ", ";
+    }
+    std::cout << "buffer write done" << std::endl;
+}
+
+void Oscillator::getSemitone()
+{
+    std::cout << "Osc is set to semitone " << this->semitone << std::endl;
+}
